refactor: remove backend deps from unit tables helpers
