+++ shared-helpers/index.ts
@@ -9,4 +9,5 @@ export * from "./src/photos"
 export * from "./src/postmarkString"
 export * from "./src/programHelpers"
 export * from "./src/stringFormatting"
+export * from "./src/summaryTables"
 export * from "./src/unitTypes"
